Partitions track slog commits in the manifest
The partition integrates the manifest and the slog together.

The slog has no metadata tracking ability. This means that the partition that sits on top of it has several responsibilities:

- tracking committed segments in the manifest
- enforcing data rolling and retention policies
- assigning unique record ids to each append

All of these responsibilities require tracking state that we cannot practically support for various different slog types. This is the primary reason the manifest exists.